+++ conn/gpio/gpio.go
@@ -11,12 +11,9 @@ package gpio
 import (
 	"errors"
 	"fmt"
-	"sort"
-	"strconv"
-	"sync"
 	"time"
 
-	"periph.io/x/periph/conn/pins"
+	"periph.io/x/periph/conn/pin"
 )
 
 // Interfaces
@@ -88,7 +85,7 @@ func (i Edge) String() string {
 //
 // It may optionally support internal pull resistor and edge based triggering.
 type PinIn interface {
-	pins.Pin
+	pin.Pin
 	// In setups a pin as an input.
 	//
 	// If WaitForEdge() is planned to be called, make sure to use one of the Edge
@@ -141,7 +138,7 @@ const (
 
 // PinOut is an output GPIO pin.
 type PinOut interface {
-	pins.Pin
+	pin.Pin
 	// Out sets a pin as output if it wasn't already and sets the initial value.
 	//
 	// After the initial call to ensure that the pin has been set as output, it
@@ -163,7 +160,7 @@ type PinOut interface {
 // It may fail at either input and or output, for example ground, vcc and other
 // similar pins.
 type PinIO interface {
-	pins.Pin
+	pin.Pin
 	In(pull Pull, edge Edge) error
 	Read() Level
 	WaitForEdge(timeout time.Duration) bool
@@ -184,180 +181,10 @@ type RealPin interface {
 	Real() PinIO // Real returns the real pin behind an Alias
 }
 
-// Registry
-
-// ByNumber returns a GPIO pin from its number.
-//
-// Returns nil in case the pin is not present.
-func ByNumber(number int) PinIO {
-	mu.Lock()
-	defer mu.Unlock()
-	return getByNumber(number)
-}
-
-// ByName returns a GPIO pin from its name.
-//
-// This can be strings like GPIO2, PB8, etc.
-//
-// This function also parses string representation of numbers, so that calling
-// with "6" will return the pin registered as number 6.
-//
-// Returns nil in case the pin is not present.
-func ByName(name string) PinIO {
-	mu.Lock()
-	defer mu.Unlock()
-	if p, ok := byName[0][name]; ok {
-		return p
-	}
-	if p, ok := byName[1][name]; ok {
-		return p
-	}
-	if p, ok := byAlias[name]; ok {
-		if p.PinIO == nil {
-			if p.PinIO = getByNumber(p.number); p.PinIO == nil {
-				return nil
-			}
-		}
-		return p
-	}
-	if i, err := strconv.Atoi(name); err == nil {
-		return getByNumber(i)
-	}
-	return nil
-}
-
-// All returns all the GPIO pins available on this host.
-//
-// The list is guaranteed to be in order of number.
-//
-// This list excludes aliases.
-//
-// This list excludes non-GPIO pins like GROUND, V3_3, etc.
-func All() []PinIO {
-	mu.Lock()
-	defer mu.Unlock()
-	out := make(pinList, 0, len(byNumber))
-	seen := make(map[int]struct{}, len(byNumber[0]))
-	// Memory-mapped pins have highest priority, include all of them.
-	for _, p := range byNumber[0] {
-		out = append(out, p)
-		seen[p.Number()] = struct{}{}
-	}
-	// Add in OS accessible pins that cannot be accessed via memory-map.
-	for _, p := range byNumber[1] {
-		if _, ok := seen[p.Number()]; !ok {
-			out = append(out, p)
-		}
-	}
-	sort.Sort(out)
-	return out
-}
-
-// Aliases returns all pin aliases.
-//
-// The list is guaranteed to be in order of aliase name.
-func Aliases() []PinIO {
-	mu.Lock()
-	defer mu.Unlock()
-	out := make(pinList, 0, len(byAlias))
-	for _, p := range byAlias {
-		// Skip aliases that were not resolved.
-		if p.PinIO == nil {
-			if p.PinIO = getByNumber(p.number); p.PinIO == nil {
-				continue
-			}
-		}
-		out = append(out, p)
-	}
-	sort.Sort(out)
-	return out
-}
-
-// Register registers a GPIO pin.
-//
-// Registering the same pin number or name twice is an error.
-//
-// `preferred` should be true when the pin being registered is exposing as much
-// functionality as possible via the underlying hardware. This is normally done
-// by accessing the CPU memory mapped registers directly.
-//
-// `preferred` should be false when the functionality is provided by the OS and
-// is limited or slower.
-//
-// The pin registered cannot implement the interface RealPin.
-func Register(p PinIO, preferred bool) error {
-	name := p.Name()
-	if len(name) == 0 {
-		return errors.New("gpio: can't register a pin with no name")
-	}
-	if _, err := strconv.Atoi(name); err == nil {
-		return fmt.Errorf("gpio: can't register a pin with a name being only a number %q", name)
-	}
-	number := p.Number()
-	if number < 0 {
-		return fmt.Errorf("gpio: can't register a pin with a negative number %d", number)
-	}
-	i := 0
-	if !preferred {
-		i = 1
-	}
-
-	mu.Lock()
-	defer mu.Unlock()
-	if orig, ok := byNumber[i][number]; ok {
-		return fmt.Errorf("gpio: can't register the same pin %d twice; had %q, registering %q", number, orig, p)
-	}
-	if orig, ok := byName[i][name]; ok {
-		return fmt.Errorf("gpio: can't register the same pin %q twice; had %q, registering %q", name, orig, p)
-	}
-	if r, ok := p.(RealPin); ok {
-		return fmt.Errorf("gpio: can't register %q, which is an aliased for %q, use RegisterAlias() instead", p, r)
-	}
-	if alias, ok := byAlias[name]; ok {
-		return fmt.Errorf("gpio: can't register %q for which an alias %q already exists", p, alias)
-	}
-	byNumber[i][number] = p
-	byName[i][name] = p
-	return nil
-}
-
-// RegisterAlias registers an alias for a GPIO pin.
-//
-// It is possible to register an alias for a pin number that itself has not
-// been registered yet.
-func RegisterAlias(alias string, number int) error {
-	if len(alias) == 0 {
-		return errors.New("gpio: can't register an alias with no name")
-	}
-	if _, err := strconv.Atoi(alias); err == nil {
-		return fmt.Errorf("gpio: can't register an alias being only a number %q", alias)
-	}
-	if number < 0 {
-		return fmt.Errorf("gpio: can't register an alias to a pin with a negative number %d", number)
-	}
-
-	mu.Lock()
-	defer mu.Unlock()
-	if orig := byAlias[alias]; orig != nil {
-		return fmt.Errorf("gpio: can't register alias %q for pin %d: it is already aliased to %q", alias, number, orig)
-	}
-	byAlias[alias] = &pinAlias{name: alias, number: number}
-	return nil
-}
-
 //
 
 // errInvalidPin is returned when trying to use INVALID.
 var errInvalidPin = errors.New("gpio: invalid pin")
-
-var (
-	mu sync.Mutex
-	// The first map is preferred pins, the second is for more limited pins,
-	// usually going through OS-provided abstraction layer.
-	byNumber = [2]map[int]PinIO{{}, {}}
-	byName   = [2]map[string]PinIO{{}, {}}
-	byAlias  = map[string]*pinAlias{}
-)
 
 func init() {
 	INVALID = invalidPin{}
@@ -406,51 +233,6 @@ func (invalidPin) Out(Level) error {
 func (invalidPin) PWM(duty int) error {
 	return errInvalidPin
 }
-
-// pinAlias implements an alias for a PinIO.
-//
-// pinAlias also implements the RealPin interface, which allows querying for
-// the real pin under the alias.
-type pinAlias struct {
-	PinIO
-	name   string
-	number int
-}
-
-// String returns the alias name along the real pin's Name() in parenthesis, if
-// known, else the real pin's number.
-func (a *pinAlias) String() string {
-	if a.PinIO == nil {
-		return fmt.Sprintf("%s(%d)", a.name, a.number)
-	}
-	return fmt.Sprintf("%s(%s)", a.name, a.PinIO.Name())
-}
-
-// Name returns the pinAlias's name.
-func (a *pinAlias) Name() string {
-	return a.name
-}
-
-// Real returns the real pin behind the alias
-func (a *pinAlias) Real() PinIO {
-	return a.PinIO
-}
-
-func getByNumber(number int) PinIO {
-	if p, ok := byNumber[0][number]; ok {
-		return p
-	}
-	if p, ok := byNumber[1][number]; ok {
-		return p
-	}
-	return nil
-}
-
-type pinList []PinIO
-
-func (p pinList) Len() int           { return len(p) }
-func (p pinList) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }
-func (p pinList) Less(i, j int) bool { return p[i].Number() < p[j].Number() }
 
 var _ PinIn = INVALID
 var _ PinOut = INVALID
