+++ conn/i2c/i2c_test.go
@@ -7,67 +7,17 @@ package i2c
 import (
 	"bytes"
 	"errors"
-	"flag"
 	"fmt"
 	"log"
-	"sort"
-	"strings"
 	"testing"
 
 	"periph.io/x/periph/conn"
 )
 
-func ExampleAll() {
-	// Enumerate all I²C buses available and the corresponding pins.
-	fmt.Print("I²C buses available:\n")
-	for _, ref := range All() {
-		fmt.Printf("- %s\n", ref.Name)
-		if ref.Number != -1 {
-			fmt.Printf("  %d\n", ref.Number)
-		}
-		if len(ref.Aliases) != 0 {
-			fmt.Printf("  %s\n", strings.Join(ref.Aliases, " "))
-		}
-
-		b, err := ref.Open()
-		if err != nil {
-			fmt.Printf("  Failed to open: %v", err)
-		}
-		if p, ok := b.(Pins); ok {
-			fmt.Printf("  SDA: %s", p.SDA())
-			fmt.Printf("  SCL: %s", p.SCL())
-		}
-		if err := b.Close(); err != nil {
-			fmt.Printf("  Failed to close: %v", err)
-		}
-	}
-}
-
-func ExampleOpen() {
-	// On linux, the following calls will likely open the same bus.
-	Open("/dev/i2c-1")
-	Open("I2C1")
-	Open("1")
-
-	// How a command line tool may let the user choose an I²C, yet default to the
-	// first bus known.
-	name := flag.String("i2c", "", "I²C bus to use")
-	flag.Parse()
-	b, err := Open(*name)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer b.Close()
-	// Use b...
-	b.Tx(23, []byte("cmd"), nil)
-}
-
 func ExampleDev() {
-	b, err := Open("")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer b.Close()
+	//b, err := i2creg.Open("")
+	//defer b.Close()
+	var b Bus
 
 	// Dev is a valid conn.Conn.
 	d := &Dev{Addr: 23, Bus: b}
@@ -81,11 +31,9 @@ func ExampleDev() {
 }
 
 func ExamplePins() {
-	b, err := Open("")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer b.Close()
+	//b, err := i2creg.Open("")
+	//defer b.Close()
+	var b Bus
 
 	// Prints out the gpio pin used.
 	if p, ok := b.(Pins); ok {
@@ -159,142 +107,6 @@ func TestDevWriteErr(t *testing.T) {
 }
 
 //
-
-func TestOpen(t *testing.T) {
-	defer reset()
-	if _, err := Open(""); err == nil {
-		t.Fatal("no bus registered")
-	}
-	if err := Register("a", []string{"x"}, 1, fakeBuser); err != nil {
-		t.Fatal(err)
-	}
-	if o, err := Open(""); o == nil || err != nil {
-		t.Fatal(o, err)
-	}
-	if o, err := Open("1"); o == nil || err != nil {
-		t.Fatal(o, err)
-	}
-	if o, err := Open("x"); o == nil || err != nil {
-		t.Fatal(o, err)
-	}
-	if o, err := Open("y"); o != nil || err == nil {
-		t.Fatal(o, err)
-	}
-}
-
-func TestDefault_NoNumber(t *testing.T) {
-	defer reset()
-	if err := Register("a", nil, -1, fakeBuser); err != nil {
-		t.Fatal(err)
-	}
-	if o, err := Open(""); o == nil || err != nil {
-		t.Fatal(o, err)
-	}
-}
-
-func TestAll(t *testing.T) {
-	defer reset()
-	if a := All(); len(a) != 0 {
-		t.Fatal(a)
-	}
-	if err := Register("a", nil, 1, fakeBuser); err != nil {
-		t.Fatal(err)
-	}
-	if err := Register("b", nil, 2, fakeBuser); err != nil {
-		t.Fatal(err)
-	}
-	if a := All(); len(a) != 2 {
-		t.Fatal(a)
-	}
-}
-
-func TestRefList(t *testing.T) {
-	l := refList{&Ref{Name: "b"}, &Ref{Name: "a"}}
-	sort.Sort(l)
-	if l[0].Name != "a" || l[1].Name != "b" {
-		t.Fatal(l)
-	}
-}
-
-func TestRegister(t *testing.T) {
-	defer reset()
-	if err := Register("a", []string{"b"}, 42, fakeBuser); err != nil {
-		t.Fatal(err)
-	}
-	if Register("a", nil, -1, fakeBuser) == nil {
-		t.Fatal("same bus name")
-	}
-	if Register("b", nil, -1, fakeBuser) == nil {
-		t.Fatal("same bus alias name")
-	}
-	if Register("c", nil, 42, fakeBuser) == nil {
-		t.Fatal("same bus number")
-	}
-	if Register("c", []string{"a"}, -1, fakeBuser) == nil {
-		t.Fatal("same bus alias")
-	}
-	if Register("c", []string{"b"}, -1, fakeBuser) == nil {
-		t.Fatal("same bus alias")
-	}
-}
-
-func TestRegister_fail(t *testing.T) {
-	defer reset()
-	if Register("a", nil, -1, nil) == nil {
-		t.Fatal("missing Opener")
-	}
-	if Register("a", nil, -2, fakeBuser) == nil {
-		t.Fatal("bad bus number")
-	}
-	if Register("", nil, 42, fakeBuser) == nil {
-		t.Fatal("missing name")
-	}
-	if Register("1", nil, 42, fakeBuser) == nil {
-		t.Fatal("numeric name")
-	}
-	if Register("a", []string{"a"}, 0, fakeBuser) == nil {
-		t.Fatal("\"a\" is already registered")
-	}
-	if Register("a", []string{""}, 0, fakeBuser) == nil {
-		t.Fatal("empty alias")
-	}
-	if Register("a", []string{"1"}, 0, fakeBuser) == nil {
-		t.Fatal("numeric alias")
-	}
-	if a := All(); len(a) != 0 {
-		t.Fatal(a)
-	}
-}
-
-func TestUnregister(t *testing.T) {
-	defer reset()
-	if Unregister("") == nil {
-		t.Fatal("unregister empty")
-	}
-	if Unregister("a") == nil {
-		t.Fatal("unregister non-existing")
-	}
-	if err := Register("a", []string{"b"}, 0, fakeBuser); err != nil {
-		t.Fatal(err)
-	}
-	if err := Unregister("a"); err != nil {
-		t.Fatal(err)
-	}
-}
-
-//
-
-func fakeBuser() (BusCloser, error) {
-	return &fakeBus{}, nil
-}
-
-func reset() {
-	mu.Lock()
-	defer mu.Unlock()
-	byName = map[string]*Ref{}
-	byNumber = map[int]*Ref{}
-	byAlias = map[string]*Ref{}
-}
 
 type fakeBus struct {
 	speed int64
