+++ conn/pin/pinreg/doc.go
@@ -2,5 +2,6 @@
 // Use of this source code is governed under the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-// Package headers describes the physical headers on a host.
-package headers
+// Package pinreg is a registry for the physical headers (made up of pins) on
+// a host.
+package pinreg
