+++ conn/pin/pinreg/pinreg.go
@@ -2,29 +2,29 @@
 // Use of this source code is governed under the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-package headers
+package pinreg
 
 import (
 	"fmt"
 	"sync"
 
 	"periph.io/x/periph/conn/gpio"
-	"periph.io/x/periph/conn/pins"
+	"periph.io/x/periph/conn/pin"
 )
 
 // All contains all the on-board headers on a micro computer.
 //
 // The map key is the header name, e.g. "P1" or "EULER" and the value is a
-// slice of slice of pins. For a 2x20 header, it's going to be a slice of
-// [20][2]pins.Pin.
-func All() map[string][][]pins.Pin {
+// slice of slice of pin.Pin. For a 2x20 header, it's going to be a slice of
+// [20][2]pin.Pin.
+func All() map[string][][]pin.Pin {
 	mu.Lock()
 	defer mu.Unlock()
-	out := make(map[string][][]pins.Pin, len(allHeaders))
+	out := make(map[string][][]pin.Pin, len(allHeaders))
 	for k, v := range allHeaders {
-		outV := make([][]pins.Pin, len(v))
+		outV := make([][]pin.Pin, len(v))
 		for i, w := range v {
-			outW := make([]pins.Pin, len(w))
+			outW := make([]pin.Pin, len(w))
 			copy(outW, w)
 			outV[i] = outW
 		}
@@ -38,7 +38,7 @@ func All() map[string][][]pins.Pin {
 // The header and the pin number. Pin numbers are 1-based.
 //
 // Returns "", 0 if not connected.
-func Position(p pins.Pin) (string, int) {
+func Position(p pin.Pin) (string, int) {
 	mu.Lock()
 	defer mu.Unlock()
 	pos, _ := byPin[realPin(p).Name()]
@@ -46,13 +46,13 @@ func Position(p pins.Pin) (string, int) {
 }
 
 // IsConnected returns true if the pin is on a header.
-func IsConnected(p pins.Pin) bool {
+func IsConnected(p pin.Pin) bool {
 	_, i := Position(p)
 	return i != 0
 }
 
 // Register registers a physical header.
-func Register(name string, allPins [][]pins.Pin) error {
+func Register(name string, allPins [][]pin.Pin) error {
 	mu.Lock()
 	defer mu.Unlock()
 	if _, ok := allHeaders[name]; ok {
@@ -85,12 +85,12 @@ type position struct {
 
 var (
 	mu         sync.Mutex
-	allHeaders = map[string][][]pins.Pin{} // every known headers as per internal lookup table
-	byPin      = map[string]position{}     // GPIO pin name to position
+	allHeaders = map[string][][]pin.Pin{} // every known headers as per internal lookup table
+	byPin      = map[string]position{}    // GPIO pin name to position
 )
 
 // realPin returns the real pin from an alias.
-func realPin(p pins.Pin) pins.Pin {
+func realPin(p pin.Pin) pin.Pin {
 	for {
 		if r, ok := p.(gpio.RealPin); ok {
 			p = r.Real()
