+++ devices/ds248x/ds248x_test.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 	"testing"
 
-	"periph.io/x/periph/conn/i2c"
+	"periph.io/x/periph/conn/i2c/i2creg"
 	"periph.io/x/periph/conn/i2c/i2ctest"
 )
 
@@ -32,7 +32,7 @@ func TestInit(t *testing.T) {
 
 func Example() {
 	// Open the I²C bus to which the DS248x is connected.
-	i2cBus, err := i2c.Open("")
+	i2cBus, err := i2creg.Open("")
 	if err != nil {
 		fmt.Println(err)
 		return
@@ -72,7 +72,7 @@ func TestRecordInit(t *testing.T) {
 	}
 	host.Init()
 
-	i2cReal, err := i2c.Open("")
+	i2cReal, err := i2creg.Open("")
 	if err != nil {
 		t.Fatal(err)
 	}
