+++ host/pine64/pine64.go
@@ -9,9 +9,9 @@ import (
 	"os"
 
 	"periph.io/x/periph"
-	"periph.io/x/periph/conn/pins"
+	"periph.io/x/periph/conn/pin"
+	"periph.io/x/periph/conn/pin/pinreg"
 	"periph.io/x/periph/host/allwinner"
-	"periph.io/x/periph/host/headers"
 )
 
 // Present returns true if running on a Pine64 board.
@@ -28,83 +28,83 @@ func Present() bool {
 
 // Pine64 specific pins.
 var (
-	VCC         = &pins.BasicPin{N: "VCC"}         //
-	IOVCC       = &pins.BasicPin{N: "IOVCC"}       // Power supply for port A
-	TEMP_SENSOR = &pins.BasicPin{N: "TEMP_SENSOR"} //
-	IR_RX       = &pins.BasicPin{N: "IR_RX"}       // IR Data Receive
-	CHARGER_LED = &pins.BasicPin{N: "CHARGER_LED"} //
-	RESET       = &pins.BasicPin{N: "RESET"}       //
-	PWR_SWITCH  = &pins.BasicPin{N: "PWR_SWITCH "} //
+	VCC         = &pin.BasicPin{N: "VCC"}         //
+	IOVCC       = &pin.BasicPin{N: "IOVCC"}       // Power supply for port A
+	TEMP_SENSOR = &pin.BasicPin{N: "TEMP_SENSOR"} //
+	IR_RX       = &pin.BasicPin{N: "IR_RX"}       // IR Data Receive
+	CHARGER_LED = &pin.BasicPin{N: "CHARGER_LED"} //
+	RESET       = &pin.BasicPin{N: "RESET"}       //
+	PWR_SWITCH  = &pin.BasicPin{N: "PWR_SWITCH "} //
 )
 
 // All the individual pins on the headers.
 var (
-	P1_1  = pins.V3_3      // max 40mA
-	P1_2  = pins.V5        // (filtered)
+	P1_1  = pin.V3_3       // max 40mA
+	P1_2  = pin.V5         // (filtered)
 	P1_3  = allwinner.PH3  //
-	P1_4  = pins.V5        // (filtered)
+	P1_4  = pin.V5         // (filtered)
 	P1_5  = allwinner.PH2  //
-	P1_6  = pins.GROUND    //
+	P1_6  = pin.GROUND     //
 	P1_7  = allwinner.PL10 //
 	P1_8  = allwinner.PB0  //
-	P1_9  = pins.GROUND    //
+	P1_9  = pin.GROUND     //
 	P1_10 = allwinner.PB1  //
 	P1_11 = allwinner.PC7  //
 	P1_12 = allwinner.PC8  //
 	P1_13 = allwinner.PH9  //
-	P1_14 = pins.GROUND    //
+	P1_14 = pin.GROUND     //
 	P1_15 = allwinner.PC12 //
 	P1_16 = allwinner.PC13 //
-	P1_17 = pins.V3_3      //
+	P1_17 = pin.V3_3       //
 	P1_18 = allwinner.PC14 //
 	P1_19 = allwinner.PC0  //
-	P1_20 = pins.GROUND    //
+	P1_20 = pin.GROUND     //
 	P1_21 = allwinner.PC1  //
 	P1_22 = allwinner.PC15 //
 	P1_23 = allwinner.PC2  //
 	P1_24 = allwinner.PC3  //
-	P1_25 = pins.GROUND    //
+	P1_25 = pin.GROUND     //
 	P1_26 = allwinner.PH7  //
 	P1_27 = allwinner.PL9  //
 	P1_28 = allwinner.PL8  //
 	P1_29 = allwinner.PH5  //
-	P1_30 = pins.GROUND    //
+	P1_30 = pin.GROUND     //
 	P1_31 = allwinner.PH6  //
 	P1_32 = allwinner.PC4  //
 	P1_33 = allwinner.PC5  //
-	P1_34 = pins.GROUND    //
+	P1_34 = pin.GROUND     //
 	P1_35 = allwinner.PC9  //
 	P1_36 = allwinner.PC6  //
 	P1_37 = allwinner.PC16 //
 	P1_38 = allwinner.PC10 //
-	P1_39 = pins.GROUND    //
+	P1_39 = pin.GROUND     //
 	P1_40 = allwinner.PC11 //
 
-	EULER_1  = pins.V3_3         //
-	EULER_2  = pins.DC_IN        //
-	EULER_3  = pins.BAT_PLUS     //
-	EULER_4  = pins.DC_IN        //
+	EULER_1  = pin.V3_3          //
+	EULER_2  = pin.DC_IN         //
+	EULER_3  = pin.BAT_PLUS      //
+	EULER_4  = pin.DC_IN         //
 	EULER_5  = TEMP_SENSOR       //
-	EULER_6  = pins.GROUND       //
+	EULER_6  = pin.GROUND        //
 	EULER_7  = IR_RX             //
-	EULER_8  = pins.V5           //
-	EULER_9  = pins.GROUND       //
+	EULER_8  = pin.V5            //
+	EULER_9  = pin.GROUND        //
 	EULER_10 = allwinner.PH8     //
 	EULER_11 = allwinner.PB3     //
 	EULER_12 = allwinner.PB4     //
 	EULER_13 = allwinner.PB5     //
-	EULER_14 = pins.GROUND       //
+	EULER_14 = pin.GROUND        //
 	EULER_15 = allwinner.PB6     //
 	EULER_16 = allwinner.PB7     //
-	EULER_17 = pins.V3_3         //
+	EULER_17 = pin.V3_3          //
 	EULER_18 = allwinner.PD4     //
 	EULER_19 = allwinner.PD2     //
-	EULER_20 = pins.GROUND       //
+	EULER_20 = pin.GROUND        //
 	EULER_21 = allwinner.PD3     //
 	EULER_22 = allwinner.PD5     //
 	EULER_23 = allwinner.PD1     //
 	EULER_24 = allwinner.PD0     //
-	EULER_25 = pins.GROUND       //
+	EULER_25 = pin.GROUND        //
 	EULER_26 = allwinner.PD6     //
 	EULER_27 = allwinner.PB2     //
 	EULER_28 = allwinner.PD7     //
@@ -112,25 +112,25 @@ var (
 	EULER_30 = allwinner.PB9     //
 	EULER_31 = allwinner.EAROUTP //
 	EULER_32 = allwinner.EAROUTN //
-	EULER_33 = pins.INVALID      //
-	EULER_34 = pins.GROUND       //
+	EULER_33 = pin.INVALID       //
+	EULER_34 = pin.GROUND        //
 
-	EXP_1  = pins.V3_3         //
+	EXP_1  = pin.V3_3          //
 	EXP_2  = allwinner.PL7     //
 	EXP_3  = CHARGER_LED       //
 	EXP_4  = RESET             //
 	EXP_5  = PWR_SWITCH        //
-	EXP_6  = pins.GROUND       //
+	EXP_6  = pin.GROUND        //
 	EXP_7  = allwinner.PB8     //
 	EXP_8  = allwinner.PB9     //
-	EXP_9  = pins.GROUND       //
+	EXP_9  = pin.GROUND        //
 	EXP_10 = allwinner.KEY_ADC //
 
-	WIFI_BT_1  = pins.GROUND        //
+	WIFI_BT_1  = pin.GROUND         //
 	WIFI_BT_2  = allwinner.PG6      //
 	WIFI_BT_3  = allwinner.PG0      //
 	WIFI_BT_4  = allwinner.PG7      //
-	WIFI_BT_5  = pins.GROUND        //
+	WIFI_BT_5  = pin.GROUND         //
 	WIFI_BT_6  = allwinner.PG8      //
 	WIFI_BT_7  = allwinner.PG1      //
 	WIFI_BT_8  = allwinner.PG9      //
@@ -143,18 +143,18 @@ var (
 	WIFI_BT_15 = allwinner.PG5      //
 	WIFI_BT_16 = allwinner.PG13     //
 	WIFI_BT_17 = allwinner.PL2      //
-	WIFI_BT_18 = pins.GROUND        //
+	WIFI_BT_18 = pin.GROUND         //
 	WIFI_BT_19 = allwinner.PL3      //
 	WIFI_BT_20 = allwinner.PL5      //
 	WIFI_BT_21 = allwinner.X32KFOUT //
 	WIFI_BT_22 = allwinner.PL5      //
-	WIFI_BT_23 = pins.GROUND        //
+	WIFI_BT_23 = pin.GROUND         //
 	WIFI_BT_24 = allwinner.PL6      //
 	WIFI_BT_25 = VCC                //
 	WIFI_BT_26 = IOVCC              //
 
-	AUDIO_LEFT  = pins.INVALID // BUG(maruel): Fix once analog is implemented.
-	AUDIO_RIGHT = pins.INVALID //
+	AUDIO_LEFT  = pin.INVALID // BUG(maruel): Fix once analog is implemented.
+	AUDIO_RIGHT = pin.INVALID //
 )
 
 //
@@ -175,7 +175,7 @@ func (d *driver) Init() (bool, error) {
 	if !Present() {
 		return false, errors.New("pine64 board not detected")
 	}
-	if err := headers.Register("P1", [][]pins.Pin{
+	if err := pinreg.Register("P1", [][]pin.Pin{
 		{P1_1, P1_2},
 		{P1_3, P1_4},
 		{P1_5, P1_6},
@@ -199,7 +199,7 @@ func (d *driver) Init() (bool, error) {
 	}); err != nil {
 		return true, err
 	}
-	if err := headers.Register("EULER", [][]pins.Pin{
+	if err := pinreg.Register("EULER", [][]pin.Pin{
 		{EULER_1, EULER_2},
 		{EULER_3, EULER_4},
 		{EULER_5, EULER_6},
@@ -221,7 +221,7 @@ func (d *driver) Init() (bool, error) {
 		return true, err
 	}
 
-	if err := headers.Register("EXP", [][]pins.Pin{
+	if err := pinreg.Register("EXP", [][]pin.Pin{
 		{EXP_1, EXP_2},
 		{EXP_3, EXP_4},
 		{EXP_5, EXP_6},
@@ -231,7 +231,7 @@ func (d *driver) Init() (bool, error) {
 		return true, err
 	}
 
-	if err := headers.Register("WIFI_BT", [][]pins.Pin{
+	if err := pinreg.Register("WIFI_BT", [][]pin.Pin{
 		{WIFI_BT_1, WIFI_BT_2},
 		{WIFI_BT_3, WIFI_BT_4},
 		{WIFI_BT_5, WIFI_BT_6},
@@ -249,7 +249,7 @@ func (d *driver) Init() (bool, error) {
 		return true, err
 	}
 
-	if err := headers.Register("AUDIO", [][]pins.Pin{
+	if err := pinreg.Register("AUDIO", [][]pin.Pin{
 		{AUDIO_LEFT},
 		{AUDIO_RIGHT},
 	}); err != nil {
